delete db_window.rs, move contents to window_service, clean up process_blobs
