+++ core/src/lib.rs
@@ -30,7 +30,6 @@ pub mod blocktree_processor;
 pub mod cluster;
 pub mod cluster_info;
 pub mod cluster_tests;
-pub mod db_window;
 pub mod entry;
 #[cfg(feature = "erasure")]
 pub mod erasure;
